widgets/requirements.txt
pygame>=2.0.0
pywebview>=4.0.0